Fix an issue where ParallelHook doesn't always continue
The fix proposed in #6 actually led to a new issue where
in specific circumstances, a Future would be left forever
pending (`never`) after resource disposal had completed.

The new solution solves the same problem that #6 did, but
in a different way. Now, instead of using a custom
transformation on top of the undocumented transform API,
we use a custom Future using the undocumented Future class.

This new solution is cleaner, clearer, and doesn't suffer
from the issue described above.